refactor: Reorganizes test fixtures
